画像取得スクリプトのmonorepo設定読み込み対応
