+++ apps/web/app/page.js
@@ -1,4 +1,3 @@
-import home_styles from "./home.css";
 import styles from "./page.module.css";
 import { Callout } from "@/app/compoent/callout.jsx";
 import { Countdown } from "@/app/compoent/countdown.jsx";
